IMP-2606 - [Backend] Schedule the automatic unsubscription notifications to users for the time between 8am and 8 pm

+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/dao/UserDao.java
@@ -39,7 +39,7 @@ public class UserDao extends JpaDaoSupport {
 				communityName).get(0);
 	}
 
-	public String getCommunityNameByUserGroup(byte userGroup) {
+	public String getCommunityNameByUserGroup(int userGroup) {
 		return ((Community) getJpaTemplate().find(
 						"select o from "
 								+ Community.class.getSimpleName()
